table component done
